We worked with that Visual Studio Package, learned about the use of Breakpoints, how to debug.
Have done lot of test, trying to check if the code works fine, and tried to reach every way of the code.

Finally the code works fine, learned about how people write whatever they want when we code something waiting a Yes/No answer.

The pins function to a variable its so usefull, we can check the value of a property every time, looks nice.

Learned too if we invest more time on debugging we gonna be better developers...